http/tests/navigation/page-cache-iframe-provisional-load.html is flaky https://bugs.webkit.org/show_bug.cgi?id=148133
Reviewed by Alexey Proskuryakov.

http/tests/navigation/page-cache-iframe-provisional-load.html would fail if
the slow load would finish before navigating away. To greatly reduce the
chances of it happening, increase the slow load delay from 100ms to 3s.

Also make the test faster by using page-cache-helper.html instead of
page-cache-helper-slow.html. The idea of using a slow helper was for the
load to potentially finish while the page was in the PageCache. However,
now that we've increased the slow load delay, this will no longer happen.

* http/tests/navigation/page-cache-iframe-provisional-load.html:
* http/tests/navigation/resources/page-cache-helper-slow.html: Removed.

Identifier: 166250@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188592 268f45cc-cd09-0410-ab3c-d52691b4dbfc