Fix a rare flakiness in svg/animations/svgPreserveAspectRatio-animation-1.html https://bugs.webkit.org/show_bug.cgi?id=148149
Patch by Said Abou-Hallawa <sabouhallawa@apple.com> on 2015-08-18
Reviewed by Tim Horton.

Do not sample the SVG animation at the middle of the animation period. At
that time, the SVG animation test script changes the value of the animated
attribute from the from-value to the to-value. It is safer to sample
immediately before and after the middle of the animation period.

* svg/animations/script-tests/svgPreserveAspectRatio-animation-1.js:
* svg/animations/svgPreserveAspectRatio-animation-1-expected.txt:

Identifier: 166271@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188613 268f45cc-cd09-0410-ab3c-d52691b4dbfc