Add null check in ImageBufferData::getData https://bugs.webkit.org/show_bug.cgi?id=148156 <rdar://problem/22337157>
Reviewed by Simon Fraser.

We're getting a number of crash reports that suggest the allocation
of the result buffer has failed, but have been unable to reproduce.
This patch adds a null check to the allocation, and logs a message
to the system console. This might avoid the crashes, and hopefully
we'll see the message.

No new tests, since we're unable to reproduce this crash.

* platform/graphics/cg/ImageBufferDataCG.cpp:
(WebCore::ImageBufferData::getData): Add a null-check and early
return.

Identifier: 166280@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188622 268f45cc-cd09-0410-ab3c-d52691b4dbfc