WK1 can re-enter layout during FrameView destruction and crash https://bugs.webkit.org/show_bug.cgi?id=148237 -and corresponding- rdar://problem/22356782
Reviewed by Zalan Bujtas.

This is a safe fix for the time being. We won't resetScrollbars() for a FrameView
that is in the PageCache because resetting scrollbars can re-enter layout for the
Frame's new FrameView, which leads to updating pages in the page cache, which
leads to a crash because we're in the process of destroying something
in the page cache. We should not be able to get into this mess at all.
* page/FrameView.cpp:
(WebCore::FrameView::~FrameView):



Identifier: 166363@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188712 268f45cc-cd09-0410-ab3c-d52691b4dbfc