[Win] Simplify menu handling code in WinLauncher https://bugs.webkit.org/show_bug.cgi?id=148461
Reviewed by Zalan Bujtas.

Revise 'ToggleMenuItem' to return a boolean value indicating if
it handled the message. Revise WndProc to use this to decide if it
should pass the message on to the default handler, rather than
duplicating the logic in both places.

* WinLauncher/Common.cpp:
(ToggleMenuItem): Return true if the menu item message was handled.
(WndProc): If 'ToggleMenuItem' did not handle the message, pass
the message tothe default handler.


Identifier: 166575@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188969 268f45cc-cd09-0410-ab3c-d52691b4dbfc