Layout Test platform/mac/fast/events/content-inset-hit-testing-in-frame.html is flaky https://bugs.webkit.org/show_bug.cgi?id=148409
Reviewed by Beth Dakin.

* WebView/WebDynamicScrollBarsView.mm:
(-[WebDynamicScrollBarsView setContentInsets:]):
Explicitly force our NSScrollView to lay out after updating
content insets. We depend on this happening synchronously (specifically,
we need our frame/bounds change callbacks to fire), because the layout
which will happen subsequently needs up-to-date information in order to
correctly a) enable scrollbars and then b) update the scroll position.

* platform/mac/fast/events/resources/iframe-to-hit-test.html:
Adjust the test so that it logs something useful if it fails
instead of complaining about not being run in WKTR/DRT.


Identifier: 166576@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@188971 268f45cc-cd09-0410-ab3c-d52691b4dbfc