[Cairo][WebGL] Upload the accelerated canvas as a texture by copying via GPU directly https://bugs.webkit.org/show_bug.cgi?id=148631
Patch by Jinyoung Hur <hur.ims@navercorp.com> on 2015-09-01
Reviewed by Dean Jackson.

When an accelerated canvas needs to be uploaded as a gl texture, it would be better to copy it to
texture directly via GPU using glCopyTexImage2D.
Note that GPU copy can not always be enabled because, with premultiplyAlpha and flipY unpack option,
it seems hard to be implemented in a way of direct GPU copy.

No new tests because there is no behavior change.

* html/canvas/WebGLRenderingContextBase.cpp:
(WebCore::WebGLRenderingContextBase::texImage2D):
* platform/graphics/ImageBuffer.cpp:
(WebCore::ImageBuffer::copyToPlatformTexture):
* platform/graphics/ImageBuffer.h:
* platform/graphics/cairo/ImageBufferCairo.cpp:
(WebCore::ImageBuffer::copyToPlatformTexture):

Identifier: 166794@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@189216 268f45cc-cd09-0410-ab3c-d52691b4dbfc