http/tests/navigation/anchor-frames-same-origin.html is flaky https://bugs.webkit.org/show_bug.cgi?id=148690
Reviewed by Zalan Bujtas.

This is a temporary workaround for Bug 148690, until we have time
to investigate why scroll propagation does not work properly if
the frame is not already layed out when we scroll to the anchor.

This change updates the test to force a layout in the grandchild
frame before we scroll to the anchor. With this change, the test
is consistently passing when run on its own or after others.

* http/tests/navigation/resources/grandchild-with-anchor.html:

Identifier: 166828@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@189250 268f45cc-cd09-0410-ab3c-d52691b4dbfc