Node.textContent = undefined should be equivalent to textContent = "" https://bugs.webkit.org/show_bug.cgi?id=148729
Reviewed by Darin Adler.

Source/WebCore:

Assigning undefined to textContent should be equivalent to assigning an empty string to it like innerHTML.
This is because textContent is defined as an DOMString? attribute in DOM4 [1] and WebIDL defines ECMAScript
undefined to be treated as null for nullable types [2].

The new behavior matches that of Firefox and Chrome.

[1] https://dom.spec.whatwg.org/#node
[2] https://heycam.github.io/webidl/#es-nullable-type

* dom/Node.idl:

LayoutTests:

Rebaselined the test now that test cases pass.

* http/tests/w3c/dom/nodes/Node-textContent-expected.txt:


Identifier: 166847@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@189269 268f45cc-cd09-0410-ab3c-d52691b4dbfc