Unreviewed, fix the module name resolution in Windows https://bugs.webkit.org/show_bug.cgi?id=148689
Attempt to fix the module name resolution in Windows.
A module name is represented as the UNIX path under the current module tests.
This fix split the module name with '/' instead of pathSeparator().

This is only utilized by the jsc.cpp for the local module tests.
So, WebKit production and JavaScriptCore framework are not affected by this change.

* jsc.cpp:
(ModuleName::startsWithRoot):
(ModuleName::ModuleName):
(resolvePath):
(GlobalObject::moduleLoaderResolve):

Identifier: 166993@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@189439 268f45cc-cd09-0410-ab3c-d52691b4dbfc